:card_file_box: remove item exemples from backend, add brand related browse and read routes
