+++ server/app/routers/api/phones/router.js
@@ -2,9 +2,16 @@ const express = require("express");
 
 const router = express.Router();
 
-const { browse, read } = require("../../../controllers/phoneActions");
+// All controllers go here
 
-router.get("/", browse);
-router.get("/:id", read);
+const phoneActions = require("../../../controllers/phoneActions");
+const brandActions = require("../../../controllers/brandActions");
 
+// All phone related routes go here
+
+router.get("/brands", brandActions.browse);
+router.get("/brands/:id", brandActions.read);
+
+router.get("/", phoneActions.browse);
+router.get("/:id", phoneActions.read);
 module.exports = router;
