+++ server/database/models/BrandRepository.js
@@ -1,0 +1,22 @@
+const AbstractRepository = require("./AbstractRepository");
+
+class BrandRepository extends AbstractRepository {
+  constructor() {
+    super({ table: "brand" });
+  }
+
+  async read(id) {
+    const [rows] = await this.database.query(
+      `select * from ${this.table} where id = ?`,
+      [id]
+    );
+    return rows[0];
+  }
+
+  async readAll() {
+    const [rows] = await this.database.query(`select * from ${this.table}`);
+    return rows;
+  }
+}
+
+module.exports = BrandRepository;
