+++ server/database/tables.js
@@ -1,6 +1,6 @@
 // Import the repository modules responsible for handling data operations on the tables
-const ItemRepository = require("./models/ItemRepository");
 const PhoneRepository = require("./models/PhoneRepository");
+const BrandRepository = require("./models/BrandRepository");
 
 // Create an empty object to hold data repositories for different tables
 const tables = {};
@@ -10,8 +10,8 @@ const tables = {};
 /* ************************************************************************* */
 
 // Register each repository as data access point for its table
-tables.item = new ItemRepository();
 tables.phone = new PhoneRepository();
+tables.brand = new BrandRepository();
 
 /* ************************************************************************* */
